+++ package.json
@@ -12,6 +12,8 @@
     "react-copy-to-clipboard": "^5.0.1",
     "react-dom": "^16.2.0",
     "react-redux": "^5.0.7",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.1",
     "redux": "^3.7.2",
     "redux-little-router": "^15.0.0",
@@ -27,7 +29,7 @@
     "precommit": "lint-staged"
   },
   "lint-staged": {
-    "*.{js,json,css,md}": ["prettier --write", "git add"]
+    "*.{js,json,css,md,less}": ["prettier --write", "git add"]
   },
   "devDependencies": {
     "babel-plugin-import": "^1.6.5",
