+++ src/console/index.js
@@ -1,8 +1,35 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-// import registerServiceWorker from './registerServiceWorker';
+import React, { Component } from "react";
+import logo from "./logo.svg";
+import "./App.css";
+import Normandy from "normandy/index.js";
+import { BrowserRouter, Link } from "react-router-dom";
+import { Route } from "react-router";
 
-ReactDOM.render(<App />, document.getElementById('root'));
-// registerServiceWorker();
+class App extends Component {
+  render() {
+    return (
+      <BrowserRouter>
+        <div className="App">
+          <header className="App-header">
+            <img src={logo} className="App-logo" alt="logo" />
+            <h1 className="App-title">Welcome to Delivery Console</h1>
+          </header>
+          <div>
+            <Route
+              path="/"
+              component={() => (
+                <div>
+                  <h1>Delivery Console</h1>
+                  <Link to="/normandy/">Normandy</Link>
+                </div>
+              )}
+            />
+            <Route path="/normandy/" component={Normandy} />
+          </div>
+        </div>
+      </BrowserRouter>
+    );
+  }
+}
+
+export default App;
