+++ src/index.js
@@ -1,7 +1,7 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import DevConsoleApp from './console/App';
+import React from "react";
+import ReactDOM from "react-dom";
+import DevConsoleApp from "./console";
 // import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<DevConsoleApp />, document.getElementById('root'));
+ReactDOM.render(<DevConsoleApp />, document.getElementById("root"));
 // registerServiceWorker();
